requirements.txt
timm
pywaggle[all]
huggingface_hub
ultralytics
torch==2.1.2
transformers==4.36.2
accelerate==0.26.1
flash-attn
pillow>=9.0.0
numpy>=1.21.0
requests>=2.25.0
mcp>=1.0.0
ollama>=0.3.0
fastmcp>=0.9.0
sage-data-client
pandas>=2.0.0
httpx>=0.25.0
asyncio-mqtt
pytest>=7.0.0
pytest-asyncio>=0.21.0